requirements.txt
blinker
click
Flask
Flask-SQLAlchemy
flask_restful
greenlet
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
psycopg2-binary
SQLAlchemy
typing_extensions
Werkzeug
zipp
pillow
